Compare string-library-based palindrome determination to stack-based
Compare two methods of determining whether the character data stored
in a singly-linked list represents a palindrome.  The first approach,
suggested by Rizavan Sipai, employs two pointers in a list to find its
middle and populate a stack. The data popped from the stack is then
compared to that in the second half of the list.  The second approach
represented in this patch relies on "man string" functions to read the
list into two characters arrays representing the first and second
halves of the list.  One array is reversed as it is created, meaning
that a single strcmp() is needed to find the answer.  The second,
character-array approach is consistently about 15% faster than the
first, stack-based one.  A countervailing advantage of the stack-based
approach is that it is easily adaptable to other data types besides
characters.