cdecl: provide one passing unit test
cdecl.c was minimally altered so as to be testable since passing stdin as a
FILE* appeared to be illegal.  Neither fdopen() nor freopen() worked with stdin.

There were several other tricky points, namely remember to flush the output
file, to seek to its beginning, and to include an explicit newline, since stdin
always has one.