cdecl: remove enumerated states in favor of boolean properties
The preconditions which are a declaration must satisfy are inclusion of exactly
one identifier, one type and one terminator.  The single-character terminator
must come at the end of the declaration, so the input sanitizer can check for
it.  The type and the identifier are extracted by the tokenizer, so have it set
a flag upon finding them.