cdecl: add support for arrays
Since the only valid place for the array delimiters is immediately following the
identifier, check upon detection of an identifier if the next character is '['.
Ignore the size parameter, if any.

Remove have_delimiter flag and replace it with is_array.  Structs, unions and
function pointers can possibly be handled the same way.