cdecl: add support for array-length declaration
Remove the legacy process_array().  If the tokenizer finds '[', check it is
followed ']', and if not, mark the identifier invalid.  If both opening and
closing brackets are present and what lies between them is digits, add them to
the output.  The program will not yet detect malformed input where the digits
are followed by alphabetic characters after the digits, just as it will not
detect nonsense in an initialization which appears after '='.