cdecl: fix failure broken recognition of stdint types
The restoration of stdint-type processing was surprisingly complex given the new
necessity of distinguishing numbers in array lengths versus those in the types.
Further add logic to reflect that each expression has only one type.   Thus
detection of numbers is ill-formed if a type has already been found and the
expression is not an array.  Consequently, refactor the tokennizer a bit and
provide many new test cases to cover the combinatorics.