cdecl: add support for multidimensional arrays
Maintain pop_stack() as a simple-ish pretty-printer which does not peek inside
the stack, replace the simple is_array bool suitable for vectors with 3 new
parser-state variables, two to track the number of array dimensions and
specified lengths, and one to track whether the last length is specified.  The
new bool is necessary because pop_stack() modifies the array_lengths parameter
but not array_dimensions.

Increase the sophistcation of tokenization and error detection in order to
account for arbitrarily many dimensions and mismatched array delimiters.  Choose
'?' to indicate an unspecified dimension except in the case of a simple vector.

Additional complexity ensued due to the need to move the identifier from the
stack top to below all of the array lengths.  Also, reorder array lengths so
that the output mentions them in the usual row-major order.  Add new unit tests
to ensure that tokenization, stack-loading and output printing all work
properly.