cdecl: replace free with __cleanup__ attribute helpers
Apparently clang and gcc both have long supported a __cleanup__ attribute which
allows designation of a resource-release or -destruction function which upon
exit of a scope.  The syntax of the cleanup function is initially a bit puzzling
because the runtime passes a pointer to the object to be released, not the
object itself.   Thus a cleanup function which is bound to a void* pointer
allocation has a signature with a void* pointer parameter, but the parameter is
actually a void**.

The cleanup function in the code is copied from systemd.  With it in place,
valgrind confirms that all allocated memory is released:

$ valgrind ./cdecl-valgrind "static int* const val[42][11];"
==828203== Memcheck, a memory error detector
==828203== Copyright (C) 2002-2024, and GNU GPL'd, by Julian Seward et al.
==828203== Using Valgrind-3.24.0 and LibVEX; rerun with -h for copyright info
==828203== Command: ./cdecl-valgrind static\ int*\ const\ val[42][11];
==828203==
val is an array of 42x11 const pointer(s) to static int

==828203==
==828203== HEAP SUMMARY:
==828203==     in use at exit: 0 bytes in 0 blocks
==828203==   total heap usage: 36 allocs, 36 frees, 1,670 bytes allocated
==828203==
==828203== All heap blocks were freed -- no leaks are possible
==828203==
==828203== For lists of detected and suppressed errors, rerun with: -s
==828203== ERROR SUMMARY: 0 errors from 0 contexts (suppressed: 0 from 0)
[alison@schallkreis C-Exercises (master)]$ git status