cdecl: support minimal function declarations with a single parameter
Handle functions with zero or one parameters.  The approach is to spawn a new
parser for each additional function parameter and to save the parsers in a
linked list while loading the stack.  In the output stage, print the description
from each of the parsers in turn.  The parser_props struct now has separate
is_function and has_function_params members since there's no need to check for
the latter without the former.   has_function_params must be checked together
with the existence test for parser->next since a linked list of parsers will
also support processing of unions and struct declarations.

The syntax of the output description still needs some work.  For multiple
parameters, the code will need to support comma-separated items and,
potentially, parameters which are themselves function pointers.