cdecl: minimal support for multiple function parameters
Handle the case where the declaration describes a function with multiple
parameters.  For each function parameter, spawn a subsidiary parser and link it
into a list whose head is the top-level one with the function name as
identifier.  Pop the stacks of the subsidiary parsers from the "type" case of
pop_stack().  That choice is admittedly a bit awkward, and triggers a
use-after-free if the top-level type is preceded by a qualifier.  The code
also fails when extra whitespace is added after a parameter-delimiting comma, or
if an extra comma is insert before ')'.  Clearly there need to be more unit
tests.  Nonetheless, the current version constitutes a major rewrite of the
one-parameter version, and is working for the simplest cases.